create a commons.js file
